Karnap
In this project a required model was needed to predict the prices of cars with the available independent variables. The management needs to understand how exactly the prices vary with the independent variables. They needed to know how they can manipulate the design of the cars, the business strategy etc to meet certain price levels. Further, the model will be a good way for management to understand the pricing dynamics of a new market, A Linear Model Regressor was use in this Analysis to predict the prices of cars to enable the company that will need to start up the production of cars in Nigeria to know the estimated prices of cars...